% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_DCLEAR.R
\name{DivideMut_DCLEAR}
\alias{DivideMut_DCLEAR}
\title{Divide muts using DCLEAR-kmer}
\usage{
DivideMut_DCLEAR(X)
}
\arguments{
\item{muts}{lineage barcode matrix}
}
\description{
Divide muts using DCLEAR-kmer
}
