% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild.R
\name{FindExpTree}
\alias{FindExpTree}
\title{LinRace main function: asymmetric division based Neighbor Joining}
\usage{
FindExpTree(states, labels, state_lineages, muts, newick_lookup, maxIter)
}
\arguments{
\item{states}{cell states of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{muts}{lineage barcode matrix}

\item{X}{input distance matrix}
}
\description{
LinRace main function: asymmetric division based Neighbor Joining
}
