% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_alter.R
\name{FindExpTree_Dif_Newick}
\alias{FindExpTree_Dif_Newick}
\title{Finding the best tree structure based on local search/ Fast Newick implementation}
\usage{
FindExpTree_Dif_Newick(
  states,
  labels,
  state_lineages,
  muts,
  prob_f,
  newick_lookup = NULL,
  maxIter = 500
)
}
\arguments{
\item{states}{the states of the gene expressions of cells}

\item{labels}{the labels of cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{muts}{lineage barcode data of cells}

\item{prob_f}{transition probabilities between cells}

\item{newick_lookup}{lookup table for internal nodes}

\item{maxIter}{maximum number of iterations of local search}
}
\description{
Finding the best tree structure based on local search/ Fast Newick implementation
}
