% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostSearch.R
\name{FindPairs}
\alias{FindPairs}
\title{Finding merging pairs based on barcodes}
\usage{
FindPairs(tree_backbone, dt)
}
\arguments{
\item{tree_backbone}{neighbor joining tree based on unique barcodes}

\item{dt}{cellid lookup table}
}
\description{
Finding merging pairs based on barcodes
}
