% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_Cas.R
\name{Generate_Mut_Unique}
\alias{Generate_Mut_Unique}
\title{Generate unique barcode data for cas hybrid runs}
\usage{
Generate_Mut_Unique(mut_dir)
}
\arguments{
\item{mut_dir}{directory to lineage barcode matrix}
}
\description{
Generate unique barcode data for cas hybrid runs
}
