% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_functions.R
\name{LikelihoodCal2}
\alias{LikelihoodCal2}
\title{Calculate combined likelihood (alternative)}
\usage{
LikelihoodCal2(
  tree,
  muts,
  cell_state_labels,
  state_lineages,
  transfer_prob = NULL,
  state_shift_prob = NULL,
  lambda = 1
)
}
\arguments{
\item{tree}{input tree}

\item{muts}{lineage barcode matrix}

\item{cell_state_labels}{cell states of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{transfer_prob}{barcode transition probabilities}

\item{state_shift_prob}{state shift probabilities}

\item{lambda}{the parameter that balances the importance of the two likelihoods}
}
\description{
This function computes the combined likelihood of a tree based on the lineage barcode and cell state data using a different way of calculating state shift probabilities
}
