% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodCal_Dif.R
\name{LikelihoodCal_Dif}
\alias{LikelihoodCal_Dif}
\title{Calculate expression likelihood with diffusion loss}
\usage{
LikelihoodCal_Dif(tree, counts, cell_state_labels, newick_lookup)
}
\arguments{
\item{tree}{input tree}

\item{counts}{gene expression data}

\item{cell_state_labels}{cell states of single cells}

\item{newick_lookup}{lookup table for internal nodes}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}
}
\description{
This function computes the expression likelihood of a tree based on cell gene expression data
}
