% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostSearch.R
\name{MergeDT}
\alias{MergeDT}
\title{Finding merging pairs based on barcodes}
\usage{
MergeDT(tree_backbone, df_pairs, dt)
}
\arguments{
\item{tree_backbone}{neighbor joining tree based on unique barcodes}

\item{df_pairs}{pairs found to be merged}

\item{dt}{cellid lookup table}
}
\description{
Finding merging pairs based on barcodes
}
