% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild.R
\name{NJ_asym}
\alias{NJ_asym}
\title{LinRace main function: asymmetric division based Neighbor Joining}
\usage{
NJ_asym(X, muts, states, state_lineages)
}
\arguments{
\item{X}{input distance matrix}

\item{muts}{lineage barcode matrix}

\item{states}{cell states of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}
}
\description{
LinRace main function: asymmetric division based Neighbor Joining
}
