% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_alter.R
\name{NJ_asym_alter}
\alias{NJ_asym_alter}
\title{LinRace main function(alter): asymmetric division based Neighbor Joining}
\usage{
NJ_asym_alter(muts, states, state_lineages, max_Iter = 200)
}
\arguments{
\item{muts}{lineage barcode matrix}

\item{states}{cell states of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{max_Iter}{the maximum iterations for local search}
}
\description{
LinRace main function(alter): asymmetric division based Neighbor Joining
}
