% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_DCLEAR.R
\name{NJ_asym_DCLEAR}
\alias{NJ_asym_DCLEAR}
\title{LinRace main function with DCLEAR: asymmetric division based DCLEAR-kmer}
\usage{
NJ_asym_DCLEAR(
  muts,
  states,
  counts,
  state_lineages,
  max_Iter = 200,
  lambda1 = 10,
  lambda2 = 1
)
}
\arguments{
\item{muts}{lineage barcode matrix}

\item{states}{cell states of single cells}

\item{counts}{gene expression data of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{max_Iter}{the maximum iterations for local search}

\item{lambda1}{hyperparameter for asymmetric division likelihood}

\item{lambda2}{hyperparameter for neighbor distance likelihood}
}
\description{
LinRace main function with DCLEAR: asymmetric division based DCLEAR-kmer
}
