% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_functions.R
\name{TreeLC2}
\alias{TreeLC2}
\title{Propose a new tree using random Subtree Swapping (rSS)}
\usage{
TreeLC2(tree, prob_internal = NULL)
}
\arguments{
\item{tree}{input tree}

\item{prob_internal}{the probability to select an internal node}
}
\description{
Propose a new tree using random Subtree Swapping (rSS)
}
